use double whenever possible

+++ src/me/kstep/ucalc/CosOp.java
@@ -4,6 +4,6 @@ class CosOp extends UOperation {
     public int arity() { return 1; }
     public CharSequence name() { return "cos"; }
     public void apply(UStack stack) {
-        stack.push(Math.cos(stack.pop().floatValue()));
+        stack.push(Math.cos(stack.pop().doubleValue()));
     }
 }
