+++ src/me/kstep/ucalc/PowerOp.java
@@ -4,8 +4,8 @@ class PowerOp extends UOperation {
     public int arity() { return 2; }
     public CharSequence name() { return "y^x"; }
     public void apply(UStack stack) {
-        float x = stack.pop().floatValue();
-        float y = stack.pop().floatValue();
+        double x = stack.pop().doubleValue();
+        double y = stack.pop().doubleValue();
         stack.push(Math.pow(y, x));
     }
 }
