+++ src/me/kstep/ucalc/SquereOp.java
@@ -4,7 +4,7 @@ class SquereOp extends UOperation {
     public int arity() { return 1; }
     public CharSequence name() { return "x²"; }
     public void apply(UStack stack) {
-        Float arg = stack.pop().floatValue();
+        double arg = stack.pop().doubleValue();
         stack.push(arg * arg);
     }
 }
