+++ src/me/kstep/ucalc/SubstructOp.java
@@ -4,6 +4,6 @@ class SubstructOp extends UOperation {
     public int arity() { return 2; }
     public CharSequence name() { return "−"; }
     public void apply(UStack stack) {
-        stack.push(-stack.pop().floatValue() + stack.pop().floatValue());
+        stack.push(-stack.pop().doubleValue() + stack.pop().doubleValue());
     }
 }
