+++ src/me/kstep/ucalc/TanOp.java
@@ -4,6 +4,6 @@ class TanOp extends UOperation {
     public int arity() { return 1; }
     public CharSequence name() { return "tan"; }
     public void apply(UStack stack) {
-        stack.push(Math.tan(stack.pop().floatValue()));
+        stack.push(Math.tan(stack.pop().doubleValue()));
     }
 }
